:truck: move src/components/magazine/list to scroll directory
