+++ front/src/components/magazine/list/scroll/Episode.jsx
