+++ front/src/components/magazine/list/scroll/Magazine.jsx
