+++ front/src/components/magazine/list/scroll/MagazinePager.jsx
