+++ front/src/components/magazine/list/scroll/Page.jsx
