+++ front/src/components/magazine/list/scroll/placeholder.jsx
